fix(llma): rename byok provider gemini to google
