+++ products/llm_analytics/backend/api/evaluations.py
@@ -20,16 +20,21 @@ from posthog.rbac.access_control_api_mixin import AccessControlViewSetMixin
 from ..models.evaluation_configs import validate_evaluation_configs
 from ..models.evaluations import Evaluation
 from ..models.model_configuration import LLMModelConfiguration
-from ..models.provider_keys import LLMProvider, LLMProviderKey
+from ..models.provider_keys import LLMProvider, LLMProviderKey, canonicalize_llm_provider
 from .metrics import llma_track_latency
 
 logger = structlog.get_logger(__name__)
 
 
+class LLMProviderChoiceField(serializers.ChoiceField):
+    def to_internal_value(self, data):
+        return super().to_internal_value(canonicalize_llm_provider(str(data)))
+
+
 class ModelConfigurationSerializer(serializers.Serializer):
     """Nested serializer for model configuration."""
 
-    provider = serializers.ChoiceField(choices=LLMProvider.choices)
+    provider = LLMProviderChoiceField(choices=LLMProvider.choices)
     model = serializers.CharField(max_length=100)
     provider_key_id = serializers.UUIDField(required=False, allow_null=True)
     provider_key_name = serializers.SerializerMethodField(read_only=True)
