+++ products/llm_analytics/backend/llm/config.py
@@ -11,14 +11,16 @@ def get_eval_config(provider: str) -> ProviderConfig | None:
     """Get eval-specific provider config with fallback to general settings."""
     from django.conf import settings
 
-    match provider:
+    canonical_provider = "google" if provider == "gemini" else provider
+
+    match canonical_provider:
         case "openai":
             api_key = settings.LLMA_EVAL_OPENAI_API_KEY or settings.OPENAI_API_KEY
             base_url = settings.LLMA_EVAL_OPENAI_BASE_URL or settings.OPENAI_BASE_URL
         case "anthropic":
             api_key = settings.LLMA_EVAL_ANTHROPIC_API_KEY or settings.ANTHROPIC_API_KEY
             base_url = None
-        case "gemini":
+        case "google":
             api_key = settings.LLMA_EVAL_GEMINI_API_KEY or settings.GEMINI_API_KEY
             base_url = None
         case _:
