+++ products/llm_analytics/backend/llm/providers/base.py
@@ -14,7 +14,7 @@ class Provider(Protocol):
 
     @property
     def name(self) -> str:
-        """Provider name (e.g., 'openai', 'anthropic', 'gemini')"""
+        """Provider name (e.g., 'openai', 'anthropic', 'google')"""
         ...
 
     def complete(
